# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment App API",
    "version": "1.0.0",
    "description": "Backend API documentation for authentication, portfolio analytics, reporting, and admin/security endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:4040",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/deposits/manual": {
      "post": {
        "summary": "Create a manual bank-transfer deposit request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "transfer_reference", "idempotency_key"],
                "properties": {
                  "amount": { "type": "number", "minimum": 0.01 },
                  "currency": { "type": "string", "enum": ["USD"], "default": "USD" },
                  "transfer_reference": { "type": "string", "minLength": 4, "maxLength": 128 },
                  "idempotency_key": { "type": "string", "minLength": 6, "maxLength": 128 },
                  "note": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Deposit request created as pending" },
          "400": { "description": "Validation error" },
          "403": { "description": "Manual deposits are disabled by rollout policy" }
        }
      }
    },
...
    "/api/v1/withdrawals/manual": {
      "post": {
        "summary": "Create a manual withdrawal request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "destination_reference", "idempotency_key"],
                "properties": {
                  "amount": { "type": "number", "minimum": 0.01 },
                  "currency": { "type": "string", "enum": ["USD"], "default": "USD" },
                  "destination_reference": { "type": "string", "minLength": 4, "maxLength": 128 },
                  "idempotency_key": { "type": "string", "minLength": 6, "maxLength": 128 },
                  "note": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Withdrawal request created as pending" },
          "400": { "description": "Validation error" },
          "403": { "description": "Manual withdrawals are disabled by rollout policy" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}